+++ cse.h
@@ -3,7 +3,7 @@
 
 #include <CSE/core.h>
 #include <CSE/systems/application.h>
-
+#include <CSE/systems/resourcemanager.h>
 #include <CSE/presets/units.h>
 
 #endif